requirements.txt
typer>=0.12.3
requests>=2.32.0
Pillow>=10.4.0
PyYAML>=6.0.2
pytest>=8.3.2
